pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "axiom"
version = "0.1.0"
description = "AXIOM — Agentic literature intelligence for researchers. Search, synthesise, detect conflicts, and surface white spaces across academic fields."
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "MD Tanvir Anjum", email = "contact@voidstudio.tech"}
]
keywords = [
    "research", "literature-review", "academic", "agent", "ai",
    "knowledge-graph", "semantic-scholar", "arxiv", "nlp", "rag"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.25.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
ollama = ["ollama>=0.1.0"]
openai = ["openai>=1.0.0"]
openrouter = ["openai>=1.0.0"]
all = ["ollama>=0.1.0", "openai>=1.0.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
]

[project.scripts]
axiom = "axiom.cli:app"

[project.urls]
Homepage = "https://github.com/tanvirpasha21/axiom-literature-agent"
Documentation = "https://github.com/tanvirpasha21/axiom-literature-agent#readme"
Repository = "https://github.com/tanvirpasha21/axiom-literature-agent"
Issues = "https://github.com/tanvirpasha21/axiom-literature-agent/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["axiom*"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
